backend/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.26
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# API Clients
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Bioinformatics and Cheminformatics
biopython>=1.83
rdkit>=2023.9.1
pymol-open-source>=3.0.0  # For structure visualization

# Structure Processing
mdanalysis>=2.6.0
biotite>=0.38.0

# Machine Learning (for property predictions)
torch>=2.0.0
transformers>=4.35.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
py3Dmol>=2.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0  # For pretty console output
loguru>=0.7.0  # Alternative logging

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# API Framework (for standalone mode)
fastapi>=0.104.0
uvicorn>=0.24.0

# Job Queue and Distributed Computing
celery>=5.3.0
redis>=5.0.0

# SSH and HPC connectivity
paramiko>=3.0.0
